style: rename app.ts to spa.ts
